requirements.txt
selenium
webdriver-manager==3.2.2
allure-pytest==2.8.6
allure-python-commons==2.8.6
pytest==5.3.2
pytest-html==2.0.0
pytest-xdist==v1.31.0
pytest-ordering==0.6
selenium-page-factory==2.0
python-dotenv==0.10.3